Got typeof working for struct fields
Some other stuff snuck in there.
