+++ src/bubba/compiler/expression/field.rs
@@ -27,7 +27,6 @@ pub(in crate::bubba::compiler) fn compile_field_access(
 
     let s = field.woog_struct;
     let s = lu_dog.exhume_woog_struct(&s).unwrap();
-    dbg!(s);
 
     // This is the expression upon which we access the field
     let expr = lu_dog.exhume_expression(&field.expression).unwrap();
