+++ src/dwarf/extruder/expression.rs
@@ -3,4 +3,5 @@ pub(super) mod and;
 pub(super) mod expr_as;
 pub(super) mod method_call;
 pub(super) mod static_method_call;
+pub(super) mod struct_expr;
 pub(super) mod unit_enum;
