+++ src/dwarf/extruder/expression/addition.rs
@@ -18,10 +18,9 @@ use crate::{
 
 // Let's just say that I don't get this lint. The docs say you have to box it
 // first, but what about when it's already boxed? I don't get it.
-#[allow(clippy::borrowed_box)]
 pub fn inter(
-    lhs_p: &Box<(ParserExpression, Range<usize>)>,
-    rhs_p: &Box<(ParserExpression, Range<usize>)>,
+    lhs_p: Box<(ParserExpression, Range<usize>)>,
+    rhs_p: Box<(ParserExpression, Range<usize>)>,
     span: RefType<Span>,
     block: &RefType<Block>,
     context: &mut Context,
