+++ src/dwarf/extruder/expression/static_method_call.rs
@@ -371,7 +371,7 @@ pub fn inter(
             lu_dog.exhume_enumeration(&woog_enum).unwrap()
         } else if lu_dog.exhume_enumeration_id_by_name(no_generics).is_some() {
             let span = s_read!(span).start as usize..s_read!(span).end as usize;
-            create_generic_enum(&type_name, no_generics, span, context, lu_dog)?.0
+            create_generic_enum(&type_name, no_generics, lu_dog)?.0
         } else {
             let span = s_read!(span).start as usize..s_read!(span).end as usize;
             return Err(vec![DwarfError::ObjectNameNotFound {
@@ -648,8 +648,7 @@ fn inter_field(
 
                     let base_name = base_path.clone() + base_name;
 
-                    let (new_enum, _) =
-                        create_generic_enum(&type_name, &base_name, span, context, lu_dog)?;
+                    let (new_enum, _) = create_generic_enum(&type_name, &base_name, lu_dog)?;
 
                     (new_enum, expr)
                 }
