+++ tests/vm_harness/types/generics/workout.ore
